backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6
supabase==2.14.0
transformers==4.36.0
torch==2.2.2
spacy==3.7.4
python-dateutil==2.9.0
numpy==1.26.4
scikit-learn==1.4.0
